requirements.txt
iterative_stratification==0.1.6
librosa==0.9.2
lightfm==1.17
ml_pyxis==0.4.dev0
numpy==1.19.5
pandas==1.1.5
pyxis==0.2.1
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.5.4
torch==1.10.1+cu111
torchmetrics==0.8.2
torchvision==0.11.2+cu111
tqdm==4.64.1
